[ARM] msm: add single-wire serial bus interface (SSBI) driver
SSBI is the Qualcomm proprietary single-wire serial bus interface used to
connect the MSM devices to the PMIC and Audio codecs.

Since SSBI only supports a single slave, the driver get the name of the slave
device passed in from the board file through the master device's platform data.

SSBI registers pretty early (postcore), so that the PMIC can come up before the
board init. This is useful if the board init requires the use of gpios
that are connected through the PMIC.

Change-Id: I07016241a0652ee93aafa855a760b8f99ad245a8
Signed-off-by: Dima Zavin <dima@android.com>